requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2